Modified annotation behaviour, to allow annotations to be made at any time. When the user requests an annotation to be made, a new fix is requested from the logging service. The annotation is then added to that fix as soon at it becomes available.
Details:

The annotation menu icon is modified while an annotation is pending. This
provides feedback that the annotation is being logged. If the user selects
the menu item again at that point, the annotation text can be edited or removed. The
latter is useful if no GPS fix is received while the user is still in
the vicinity of the location that she wanted to annotate, so that the
annotation is now irrelevant.

When logging an annotation, the limits on time and distance between points
are temporarlily relaxed. If no logging session is in progress, "single point"
logging will be activated. Otherwise, we just request a new point.