Ignore the LICENSE.md that gets copied into text, but let eleventy still process it.
The docker compose mounts LICENSE.md into text/ so that it can be processed as a standalone page. Fine. But, LICENSE.md then appears as a new file in git when working locally.
So I add it to the .gitignore. But eleventy takes .gitignore into account when processing so it ignored the LICENSE.md!
So I've got eleventy to ignore .gitignore, and use .eleventyignore instead.  I'm never going to remember this.