Add flag to custom url logging which can disable queueing requests while being offline
Issue #1025

This change is to solve the following use case:
* Logging interval is 15 seconds.
* Custom url logging is enabled.
* There is no network connection for 3 hours.
* A custom url call takes around 500 milliseconds.
* This means after the phone has network connection again, it will take
  6 minutes to send the queued requests.
* If only the current location is relevant for the user, there is no need
to send all the previous locations. This decreases the number of requests
and the delay until the current location is sent to the server.

A new switch is added to the customurlsettings page.
This is false by default. In this case everything works as before.

If the flag is set to true, old custom url log jobs tagged with 'Discardable'
will be removed from the queue when a new job is created.
This way when the network connection is restored, only one custom url log job
will be in the queue which will contain the latest location.

This change does not affect in any way the auto-send to custom url (when a
list of calls are made based on the csv file content). Jobs created for
auto send are not tagged with 'Discardable' and won't be cancelled by the
log jobs.