Use exact alarms for auto send timers
Issue #999

The use case was that after an Android version update
the auto send did not run every minute anymore as defined
in settings, but ran only every 5 minutes when the phone
was idle.

This is based on commits ee96c9bd and e9521818 where
a similar change was made with the alarms used for
the logging interval.
The related issue is #969.

With 'setAndAllowWhileIdle' the alarm might be delayed after
the specified time passed when the phone is idle, while with
'setExactAndAllowWhileIdle' the system will try to trigger the
alarm exactly at the defined time.