"Experimental: Using significant motion sensor to prevent logging, to save battery
This uses the TYPE_SIGNIFICANT_MOTION sensor. https://developer.android.com/develop/sensors-and-location/sensors/sensors_motion#sensors-motion-significant
"A significant motion is a motion that might lead to a change in the user's location; for example walking, biking, or sitting in a moving car."
I have no idea how I overlooked this for 10 years.

The significant motion detection is provided by the OS itself, so it could be completely 'off' for some devices.

Issue #223
Issue #1199
Issue #807
Issue #1208